Extraer feedback visual de terminal.js
Segundo paso de la descomposición de terminal.js (después del dedup
de audio-feedback en PR #158): updateClock(), updateIdleDate(),
setTerminalVideoState(), setIdStatusDot() y los dots de progreso de
captura facial se mueven a terminal/ui-feedback.js — mismo patrón ya
usado para mark.js (ver mark/ui-feedback.js), pero no es el mismo
módulo: los IDs del DOM y algunas clases CSS de estado difieren entre
mark.js y terminal.js (el terminal tiene reloj propio de splash idle,
por ejemplo).

Todas son funciones puramente visuales sin estado compartido con el
resto de terminal.js — no necesitaron cambios en ninguno de sus call
sites. Refactor puro, sin cambios de comportamiento.
